requirements.txt
absl-py==2.1.0
accelerate==1.0.1
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.5.2
async-timeout==5.0.1
attrs==25.1.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.1
datasets==3.1.0
dill==0.3.8
docker-pycreds==0.4.0
eval_type_backport==0.2.2
exceptiongroup==1.2.2
fastapi==0.115.11
filelock==3.16.1
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2024.9.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.38.0
google-auth-oauthlib==1.0.0
grpcio==1.70.0
h11==0.14.0
huggingface-hub==0.29.2
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
Jinja2==3.1.6
joblib==1.4.2
jsonlines==4.0.0
kiwisolver==1.4.7
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
networkx==3.1
numpy==1.21.0
oauthlib==3.2.2
omegaconf==2.3.0
packaging==24.2
pandas==2.0.3
pillow==10.4.0
platformdirs==4.3.6
propcache==0.2.0
protobuf==5.29.3
psutil==7.0.0
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytorch-crf==0.7.2
pytz==2025.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rsa==4.9
safetensors==0.5.3
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
sentencepiece==0.2.0
sentry-sdk==2.22.0
seqeval==1.2.2
setproctitle==1.3.5
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
starlette==0.44.0
sympy==1.13.3
tensorboard==2.14.0
tensorboard-data-server==0.7.2
threadpoolctl==3.5.0
tokenizers==0.20.3
torch==2.4.1
torchaudio==2.4.1
torchvision==0.19.1
tqdm==4.67.1
transformers==4.46.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.2.3
uvicorn==0.33.0
wandb==0.19.8
Werkzeug==3.0.6
xxhash==3.5.0
yarl==1.15.2
zipp==3.20.2
